+++ src/pages/Settings.tsx
@@ -1,6 +1,7 @@
 import clsx from 'clsx'
 import { render } from 'preact'
 import { useEffect, useState } from 'preact/hooks'
+import { Settings } from '../settings'
 
 export interface SettingsProps {
   mode: 'popup' | 'options'
@@ -11,6 +12,8 @@ export function SettingsPage(props: SettingsProps) {
   const [returnToLastTime, setReturnToLastTime] = useState(true)
   const [saveInterval, setSaveInterval] = useState(5)
   const [useTimestamps, setUseTimestamps] = useState(true)
+  const [completedPercent, setCompletedPercent] = useState(95)
+  const [showCompletion, setShowCompletion] = useState(true)
 
   useEffect(() => {
     chrome.storage.sync.get(
@@ -18,11 +21,15 @@ export function SettingsPage(props: SettingsProps) {
         returnToLastTime: false,
         saveInterval: 5,
         useTimestamps: true,
-      },
+        completedPercent: 95,
+        showCompletion: true,
+      } as Settings,
       (items) => {
         setReturnToLastTime(items.returnToLastTime)
         setSaveInterval(items.saveInterval)
         setUseTimestamps(items.useTimestamps)
+        setCompletedPercent(items.completedPercent)
+        setShowCompletion(items.showCompletion)
       },
     )
   }, [])
@@ -36,6 +43,12 @@ export function SettingsPage(props: SettingsProps) {
   useEffect(() => {
     chrome.storage.sync.set({ useTimestamps })
   }, [useTimestamps])
+  useEffect(() => {
+    chrome.storage.sync.set({ completedPercent })
+  }, [completedPercent])
+  useEffect(() => {
+    chrome.storage.sync.set({ showCompletion })
+  }, [showCompletion])
 
   return (
     <div className={'min-w-[350px] max-w-xl w-full mx-auto p-4'}>
@@ -55,6 +68,36 @@ export function SettingsPage(props: SettingsProps) {
             onChange={(e) => setReturnToLastTime((e.target as HTMLInputElement).checked)}
           />{' '}
           Return to last played time when returning to a video
+        </label>
+      </SettingRow>
+      <SettingRow>
+        <label className="text-base">
+          Time to mark as completed
+          <br />
+          <div className="flex items-center">
+            <input
+              type="range"
+              min={50}
+              max={100}
+              value={completedPercent}
+              step={5}
+              onInput={(e) => setCompletedPercent(parseInt((e.target as HTMLInputElement).value))}
+            />
+            <span>{completedPercent}%</span>
+          </div>
+          <p className="text-sm">
+            When video is played to this percentage, it will be marked as completed.
+          </p>
+        </label>
+      </SettingRow>
+      <SettingRow>
+        <label className="text-base">
+          <input
+            type="checkbox"
+            checked={showCompletion}
+            onChange={(e) => setShowCompletion((e.target as HTMLInputElement).checked)}
+          />{' '}
+          Show progress bar on video thumbnails
         </label>
       </SettingRow>
       <SettingRow>